greenaware/requirements.txt
requests
bootstrap
whitenoise
python-dotenv
PyJWT
stripe